+++ lib/assignments/assignment_two/profile_screen.dart
@@ -1,5 +1,8 @@
+
+import 'package:assignment_flutter/assignments/assignment_two/user_modal.dart';
+import 'package:assignment_flutter/constants.dart';
 import 'package:flutter/material.dart';
-import 'home_screen.dart'; // ensure it imports your User class
+ 
 
 class Profile extends StatelessWidget {
   final User user;
@@ -9,22 +12,40 @@ class Profile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text(user.name)),
+      appBar: AppBar(title: Flexible(child: Text(user.name))),
       body: Center(
         child: Column(
           children: [
             const SizedBox(height: 20),
-            CircleAvatar(radius: 40, child: Text(user.name[0])),
+            CircleAvatar(
+              radius: 40,
+              child: Expanded(child: Text(user.name[0])),
+            ),
             const SizedBox(height: 20),
-            Text(
-              '${user.name} ',
-              style: const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+            Flexible(
+              child: Text(
+                '${user.name} ',
+                style: const TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
             ),
-            Text(
-              'Age:  ${user.age} ',
-              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+            Flexible(
+              child: Text(
+                '$age:  ${user.age} ',
+                style: const TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
             ),
-            Text(' ${user.location} ', style: const TextStyle(fontSize: 16)),
+            Flexible(
+              child: Text(
+                ' ${user.location} ',
+                style: const TextStyle(fontSize: 16),
+              ),
+            ),
 
             const SizedBox(height: 20),
             Padding(
@@ -32,11 +53,13 @@ class Profile extends StatelessWidget {
               child: Container(
                 color: Color.fromARGB(255, 220, 245, 245),
                 padding: const EdgeInsets.all(10),
-                child: Text(
-                  '${user.description},',
-                  style: const TextStyle(
-                    fontSize: 16,
-                    backgroundColor: const Color.fromARGB(255, 220, 245, 245),
+                child: Flexible(
+                  child: Text(
+                    '${user.description},',
+                    style: const TextStyle(
+                      fontSize: 16,
+                      backgroundColor:  Color.fromARGB(255, 220, 245, 245),
+                    ),
                   ),
                 ),
               ),
@@ -48,11 +71,13 @@ class Profile extends StatelessWidget {
                   backgroundColor: const Color.fromARGB(255, 235, 237, 238),
                 ),
                 onPressed: () {
-                  print("Login Clicked");
+                  // login clicked
                 },
-                child: const Text(
-                  "Follow",
-                  style: TextStyle(color: Color.fromARGB(255, 26, 25, 25)),
+                child: const Flexible(
+                  child: Text(
+                    follow,
+                    style: TextStyle(color: Color.fromARGB(255, 26, 25, 25)),
+                  ),
                 ),
               ),
             ),
