+++ lib/assignments/assignment_two/user_modal.dart
@@ -1,0 +1,17 @@
+// lib/models/user_model.dart
+
+class User {
+  final String name;
+  final int age;
+  final String description;
+  final String location;
+  final int experiance;
+
+  User({
+    required this.name,
+    required this.age,
+    required this.description,
+    required this.location,
+    required this.experiance,
+  });
+}
