+++ lib/constants.dart
@@ -1,5 +1,47 @@
+import 'package:assignment_flutter/assignments/assignment_two/user_modal.dart';
+
 const String profilePageTitle = "Profile Page";
 const String name = "Pooja Lande";
 const String profileDescription ="Experienced and passionate front-end and cross-platform mobile developer with over 4.5+ years of expertise in building scalable, high-performance applications using React Native, React.js, and Flutter. Proven ability to design, develop, and deliver elegant UI/UX-driven mobile apps and responsive web interfaces across iOS, Android, and web platforms Skilled in integrating RESTful APIs, handling state management (Redux, Provider, Bloc), optimizing performance, writing reusable components, and ensuring pixel-perfect UI across devices.";
 const String btnText = "Follow";
+const String  homeScreen = "Home Screen";
+const String follow = "Follow";
+const String age = "Age";
 
+final List<User> users = [
+  User(
+    name: 'Pooja Lande',
+    age: 26,
+    description: 'Flutter developer & coffee lover',
+    location: "Pune",
+    experiance: 4,
+  ),
+  User(
+    name: 'Kanak Kakad',
+    age: 26,
+    description: 'Designer and runner',
+    location: "Nagpur",
+    experiance: 5,
+  ),
+  User(
+    name: 'Aarati Dike',
+    age: 26,
+    description: 'Photographer & traveler',
+    location: "Nashik",
+    experiance: 6,
+  ),
+  User(
+    name: 'Minakshi Deshmukh',
+    age: 26,
+    description: 'Mobile app developer',
+    location: "Yavatmal",
+    experiance: 3,
+  ),
+  User(
+    name: 'Aarti Dike',
+    age: 26,
+    description: 'AWS/web developer',
+    location: "Amaravati",
+    experiance: 7,
+  ),
+];
